Separated classes into files
VertetxToTriangle is now OOP, classes are separated into cpp and header files, 3d.h is gone
