+++ 3D Graph/3D Graph.vcxproj
@@ -147,13 +147,18 @@
     <Text Include="ReadMe.txt" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="3d.h" />
+    <ClInclude Include="Camera.h" />
+    <ClInclude Include="Triangle.h" />
     <ClInclude Include="render.h" />
     <ClInclude Include="stdafx.h" />
     <ClInclude Include="targetver.h" />
+    <ClInclude Include="Vector3D.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="3D Graph.cpp" />
+    <ClCompile Include="Camera.cpp" />
+    <ClCompile Include="Vector3D.cpp" />
+    <ClCompile Include="Triangle.cpp" />
     <ClCompile Include="stdafx.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">Create</PrecompiledHeader>
