+++ 3D Graph/stdafx.h
@@ -1,7 +1,7 @@
 // stdafx.h : include file for standard system include files,
 // or project specific include files that are used frequently, but
 // are changed infrequently
-//
+//AUTHOR: archit120
 
 #pragma once
